# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Controle de Estoque API",
        "description": "API para controle de estoque, movimentações e autenticação",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Leonardo Losso",
            "email": "leonardolosso.sz@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Lista todas as categorias",
                "operationId": "05296a0be702181399ad8c435ded282a",
                "responses": {
                    "200": {
                        "description": "Lista de categorias retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Cria uma nova categoria",
                "operationId": "b40e21b97a4a91e3458595c6e6b0f9d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da categoria",
                                        "type": "string",
                                        "example": "Categoria 1"
                                    },
                                    "description": {
                                        "description": "Descrição da categoria",
                                        "type": "string",
                                        "example": "Descrição breve."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação não autorizada"
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Busca uma categoria pelo ID",
                "operationId": "ab59507bc38c7f7187a35af2b3a6f706",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualiza uma categoria existente",
                "operationId": "76a62b248a8b5b20c2f5b54bfcdc0369",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da categoria",
                                        "type": "string",
                                        "example": "Categoria Atualizada"
                                    },
                                    "description": {
                                        "description": "Descrição atualizada",
                                        "type": "string",
                                        "example": "Descrição atualizada."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    },
                    "403": {
                        "description": "Ação não autorizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Exclui uma categoria",
                "operationId": "759e24e4fa1dd8c0d46c33d70e0a6e0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoria excluída com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    },
                    "403": {
                        "description": "Ação não autorizada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID da categoria",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome da categoria",
                        "type": "string",
                        "example": "Tecnologia"
                    },
                    "description": {
                        "description": "Descrição da categoria",
                        "type": "string",
                        "example": "Categoria voltada para produtos de tecnologia."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}